requirements.txt
celery==5.1.2
Django==3.2.7
djangorestframework==3.12.4
drf-renderer-xlsx==0.4.1
openpyxl==3.0.7
pandas==1.2.4
psycopg2-binary==2.9.1
redis==3.5.3
xlsxwriter==1.4.3